Support calling a local heroku-api
When trying to use a local heroku api host like
HEROKU_HOST="http://localhost:5001", the api client fails for a number
of reasons. First, protocol is not set, so it defaults to https. Second,
the port is not split out from the hostname so it attempts to lookup the
host using the portname. Third, the port is not passed in to the http
client. This patch fixes all of these issues.